refactored health system
originally, the health UI was being set in update, along with the time
scale being set to 0 in *every single update loop*. This meant that no
matter how many times it was set back to 1, the moment the time scale
went to 1, the next frame it would be set to 0 again since the health
value was 0. The current update health function is only fired on ball
kill. This lets the button hooks work as intended